Use a predictible resultAlias value for each possible Tomcat ThreadPool
http-nio and ajp-nio Thread Pools MBean's names are always suffixed with the port number. When the port number is randomly assigned (running in a PaaS like CloudFoundry), metrics will have different names every times tomcat is (re)started. This can be an issue in case you have a metric quota on the backend (hostedgraphite.com). Even in the absence of a quota, it will simply fillup the backend disk. Also the metrics will be shortlive and cannot be easilly followed in the longtime.